Remove redundant files, Update tests
