--- app/views/inventory_foods/show.html.erb
@@ -1,10 +1,0 @@
-<p style="color: green"><%= notice %></p>
-
-<%= render @inventory_food %>
-
-<div>
-  <%= link_to "Edit this inventory food", edit_inventory_food_path(@inventory_food) %> |
-  <%= link_to "Back to inventory foods", inventory_foods_path %>
-
-  <%= button_to "Destroy this inventory food", @inventory_food, method: :delete %>
-</div>
