--- app/views/recipe_foods/index.html.erb
@@ -1,14 +1,0 @@
-<p style="color: green"><%= notice %></p>
-
-<h1>Recipe foods</h1>
-
-<div id="recipe_foods">
-  <% @recipe_foods.each do |recipe_food| %>
-    <%= render recipe_food %>
-    <p>
-      <%= link_to "Show this recipe food", recipe_food %>
-    </p>
-  <% end %>
-</div>
-
-<%= link_to "New recipe food", new_recipe_food_path %>
