--- app/views/recipe_foods/show.html.erb
@@ -1,10 +1,0 @@
-<p style="color: green"><%= notice %></p>
-
-<%= render @recipe_food %>
-
-<div>
-  <%= link_to "Edit this recipe food", edit_recipe_food_path(@recipe_food) %> |
-  <%= link_to "Back to recipe foods", recipe_foods_path %>
-
-  <%= button_to "Destroy this recipe food", @recipe_food, method: :delete %>
-</div>
