--- app/views/users/_user.html.erb
@@ -1,7 +1,0 @@
-<div id="<%= dom_id user %>">
-  <p>
-    <strong>Name:</strong>
-    <%= user.name %>
-  </p>
-
-</div>
