--- app/views/users/_user.json.jbuilder
@@ -1,2 +1,0 @@
-json.extract! user, :id, :name, :email, :created_at, :updated_at
-json.url user_url(user, format: :json)
