+++ spec/features/shopping_list_generator_spec.rb
@@ -40,18 +40,6 @@ RSpec.describe 'Shopping list generator', type: :feature do
     login_user(user)
   end
 
-  def login_user(user)
-    visit '/users/sign_in'
-    fill_in 'user_email', with: user.email
-    fill_in 'user_password', with: user.password
-    click_button 'Log in'
-    wait_for_page_load
-  end
-
-  def wait_for_page_load
-    expect(page).to have_content('Signed in successfully.')
-  end
-
   describe 'Shopping list generator' do
     before(:each) do
       visit "/recipes/#{recipe.id}"
